Hand refactor for readability
